feat: add parallel-dev bin && Makefile
