+++ node/parallel-dev/src/rpc.rs
@@ -2,13 +2,13 @@
 
 use std::sync::Arc;
 
-use primitives::*;
+use primitives::{AccountId, Balance, Index};
 pub use sc_rpc_api::DenyUnsafe;
 use sp_api::ProvideRuntimeApi;
 use sp_block_builder::BlockBuilder;
 use sp_blockchain::{Error as BlockChainError, HeaderBackend, HeaderMetadata};
 use sp_transaction_pool::TransactionPool;
-use vanilla_runtime::{opaque::Block, AccountId, Index};
+use vanilla_runtime::opaque::Block;
 
 /// Full client dependencies.
 pub struct FullDeps<C, P> {
