+++ node/parallel-dev/src/service.rs
@@ -11,78 +11,56 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-use cumulus_client_consensus_relay_chain::{
-    build_relay_chain_consensus, BuildRelayChainConsensusParams,
-};
-use cumulus_client_network::build_block_announce_validator;
-use cumulus_client_service::{
-    prepare_node_config, start_collator, start_full_node, StartCollatorParams, StartFullNodeParams,
-};
-use cumulus_primitives_core::ParaId;
-use polkadot_primitives::v0::CollatorPair;
 use sc_client_api::{ExecutorProvider, RemoteBackend};
 use sc_consensus_aura::{ImportQueueParams, SlotProportion, StartAuraParams};
 use sc_executor::native_executor_instance;
 pub use sc_executor::NativeExecutor;
 use sc_finality_grandpa::SharedVoterState;
 use sc_keystore::LocalKeystore;
-use sc_service::{
-    error::Error as ServiceError, Configuration, PartialComponents, Role, TFullClient, TaskManager,
-};
-use sc_telemetry::{Telemetry, TelemetryWorker, TelemetryWorkerHandle};
+use sc_service::{error::Error as ServiceError, Configuration, TaskManager};
+use sc_telemetry::{Telemetry, TelemetryWorker};
 use sp_consensus_aura::sr25519::AuthorityPair as AuraPair;
 use sp_core::Pair;
 use sp_inherents::InherentDataProviders;
 use sp_keystore::SyncCryptoStore;
-use sp_runtime::traits::BlakeTwo256;
-use sp_trie::PrefixedMemoryDB;
-
 use std::sync::Arc;
 use std::time::Duration;
+use vanilla_runtime::{self, opaque::Block, RuntimeApi};
 
-// Native executor instance.
+// Our native executor instance.
 native_executor_instance!(
     pub Executor,
-    parallel_runtime::api::dispatch,
-    parallel_runtime::native_version,
+    vanilla_runtime::api::dispatch,
+    vanilla_runtime::native_version,
     frame_benchmarking::benchmarking::HostFunctions,
 );
 
-type ParallelBlock = parallel_runtime::opaque::Block;
-type ParallelRuntimeApi = parallel_runtime::RuntimeApi;
-type ParallelFullClient = sc_service::TFullClient<ParallelBlock, ParallelRuntimeApi, Executor>;
-type ParallelFullBackend = sc_service::TFullBackend<ParallelBlock>;
+type FullClient = sc_service::TFullClient<Block, RuntimeApi, Executor>;
+type FullBackend = sc_service::TFullBackend<Block>;
+type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
 
-type VanillaBlock = vanilla_runtime::opaque::Block;
-type VanillaRuntimeApi = vanilla_runtime::RuntimeApi;
-type VanillaFullClient = sc_service::TFullClient<VanillaBlock, VanillaRuntimeApi, Executor>;
-type VanillaFullBackend = sc_service::TFullBackend<VanillaBlock>;
-
-type FullSelectChain = sc_consensus::LongestChain<VanillaFullBackend, VanillaBlock>;
-
-pub fn new_partial_vanilla(
+pub fn new_partial(
     config: &Configuration,
 ) -> Result<
     sc_service::PartialComponents<
-        VanillaFullClient,
-        VanillaFullBackend,
+        FullClient,
+        FullBackend,
         FullSelectChain,
-        sp_consensus::DefaultImportQueue<VanillaBlock, VanillaFullClient>,
-        sc_transaction_pool::FullPool<VanillaBlock, VanillaFullClient>,
+        sp_consensus::DefaultImportQueue<Block, FullClient>,
+        sc_transaction_pool::FullPool<Block, FullClient>,
         (
             sc_consensus_aura::AuraBlockImport<
-                VanillaBlock,
-                VanillaFullClient,
+                Block,
+                FullClient,
                 sc_finality_grandpa::GrandpaBlockImport<
-                    VanillaFullBackend,
-                    VanillaBlock,
-                    VanillaFullClient,
+                    FullBackend,
+                    Block,
+                    FullClient,
                     FullSelectChain,
                 >,
                 AuraPair,
             >,
-            sc_finality_grandpa::LinkHalf<VanillaBlock, VanillaFullClient, FullSelectChain>,
+            sc_finality_grandpa::LinkHalf<Block, FullClient, FullSelectChain>,
             Option<Telemetry>,
         ),
     >,
@@ -107,16 +85,11 @@ pub fn new_partial_vanilla(
         .transpose()?;
 
     let (client, backend, keystore_container, task_manager) =
-        sc_service::new_full_parts::<VanillaBlock, VanillaRuntimeApi, Executor>(
+        sc_service::new_full_parts::<Block, RuntimeApi, Executor>(
             &config,
             telemetry.as_ref().map(|(_, telemetry)| telemetry.handle()),
         )?;
     let client = Arc::new(client);
-
-    let telemetry = telemetry.map(|(worker, telemetry)| {
-        task_manager.spawn_handle().spawn("telemetry", worker.run());
-        telemetry
-    });
 
     // PAI-NOTE: For inserting key be used in ocw
     let secret_uri = "//Alice";
@@ -143,6 +116,11 @@ pub fn new_partial_vanilla(
         key_pair.public().as_ref(),
     )
     .expect("Insert key should succeed");
+
+    let telemetry = telemetry.map(|(worker, telemetry)| {
+        task_manager.spawn_handle().spawn("telemetry", worker.run());
+        telemetry
+    });
 
     let select_chain = sc_consensus::LongestChain::new(backend.clone());
 
@@ -195,6 +173,14 @@ pub fn new_partial_vanilla(
     })
 }
 
+fn remote_keystore(_url: &String) -> Result<Arc<LocalKeystore>, &'static str> {
+    // FIXME: here would the concrete keystore be built,
+    //        must return a concrete type (NOT `LocalKeystore`) that
+    //        implements `CryptoStore` and `SyncCryptoStore`
+    Err("Remote Keystore not supported.")
+}
+
+/// Builds a new service for a full client.
 pub fn new_full(mut config: Configuration) -> Result<TaskManager, ServiceError> {
     let sc_service::PartialComponents {
         client,
@@ -206,7 +192,7 @@ pub fn new_full(mut config: Configuration) -> Result<TaskManager, ServiceError> 
         transaction_pool,
         inherent_data_providers,
         other: (block_import, grandpa_link, mut telemetry),
-    } = new_partial_vanilla(&config)?;
+    } = new_partial(&config)?;
 
     if let Some(url) = &config.keystore_remote {
         match remote_keystore(url) {
@@ -368,118 +354,6 @@ pub fn new_full(mut config: Configuration) -> Result<TaskManager, ServiceError> 
     Ok(task_manager)
 }
 
-fn remote_keystore(_url: &String) -> Result<Arc<LocalKeystore>, &'static str> {
-    // FIXME: here would the concrete keystore be built,
-    //        must return a concrete type (NOT `LocalKeystore`) that
-    //        implements `CryptoStore` and `SyncCryptoStore`
-    Err("Remote Keystore not supported.")
-}
-
-/// Starts a `ServiceBuilder` for a full service.
-///
-/// Use this macro if you don't actually need the full service, but just the builder in order to
-/// be able to perform chain operations.
-pub fn new_partial_parallel(
-    config: &Configuration,
-) -> Result<
-    PartialComponents<
-        ParallelFullClient,
-        ParallelFullBackend,
-        (),
-        sp_consensus::import_queue::BasicQueue<ParallelBlock, PrefixedMemoryDB<BlakeTwo256>>,
-        sc_transaction_pool::FullPool<
-            ParallelBlock,
-            TFullClient<ParallelBlock, ParallelRuntimeApi, Executor>,
-        >,
-        (Option<Telemetry>, Option<TelemetryWorkerHandle>),
-    >,
-    sc_service::Error,
-> {
-    let inherent_data_providers = sp_inherents::InherentDataProviders::new();
-
-    let telemetry = config
-        .telemetry_endpoints
-        .clone()
-        .filter(|x| !x.is_empty())
-        .map(|endpoints| -> Result<_, sc_telemetry::Error> {
-            let worker = TelemetryWorker::new(16)?;
-            let telemetry = worker.handle().new_telemetry(endpoints);
-            Ok((worker, telemetry))
-        })
-        .transpose()?;
-
-    let (client, backend, keystore_container, task_manager) =
-        sc_service::new_full_parts::<ParallelBlock, ParallelRuntimeApi, Executor>(
-            &config,
-            telemetry.as_ref().map(|(_, telemetry)| telemetry.handle()),
-        )?;
-    let client = Arc::new(client);
-
-    // PAI-NOTE: For inserting key be used in ocw
-    let secret_uri = "//Alice";
-    let key_pair = parallel_runtime::pallet_ocw_oracle::crypto::Pair::from_string(secret_uri, None)
-        .expect("Generates key pair");
-    let keystore = keystore_container.sync_keystore();
-    SyncCryptoStore::insert_unknown(
-        &*keystore,
-        parallel_runtime::pallet_ocw_oracle::KEY_TYPE,
-        secret_uri,
-        key_pair.public().as_ref(),
-    )
-    .expect("Insert key should succeed");
-
-    //TODO please refer to https://github.com/parallel-finance/parallel/issues/47
-    let secret_uri = "//Pool";
-    let key_pair = parallel_runtime::pallet_liquidate::crypto::Pair::from_string(secret_uri, None)
-        .expect("Generates key pair");
-    let keystore = keystore_container.sync_keystore();
-    SyncCryptoStore::insert_unknown(
-        &*keystore,
-        parallel_runtime::pallet_liquidate::KEY_TYPE,
-        secret_uri,
-        key_pair.public().as_ref(),
-    )
-    .expect("Insert key should succeed");
-
-    let telemetry_worker_handle = telemetry.as_ref().map(|(worker, _)| worker.handle());
-
-    let telemetry = telemetry.map(|(worker, telemetry)| {
-        task_manager.spawn_handle().spawn("telemetry", worker.run());
-        telemetry
-    });
-
-    let registry = config.prometheus_registry();
-
-    let transaction_pool = sc_transaction_pool::BasicPool::new_full(
-        config.transaction_pool.clone(),
-        config.role.is_authority().into(),
-        config.prometheus_registry(),
-        task_manager.spawn_handle(),
-        client.clone(),
-    );
-
-    let import_queue = cumulus_client_consensus_relay_chain::import_queue(
-        client.clone(),
-        client.clone(),
-        inherent_data_providers.clone(),
-        &task_manager.spawn_essential_handle(),
-        registry,
-    )?;
-
-    let params = PartialComponents {
-        backend,
-        client,
-        import_queue,
-        keystore_container,
-        task_manager,
-        transaction_pool,
-        inherent_data_providers,
-        select_chain: (),
-        other: (telemetry, telemetry_worker_handle),
-    };
-
-    Ok(params)
-}
 /// Builds a new service for a light client.
 pub fn new_light(mut config: Configuration) -> Result<TaskManager, ServiceError> {
     let telemetry = config
@@ -494,7 +368,7 @@ pub fn new_light(mut config: Configuration) -> Result<TaskManager, ServiceError>
         .transpose()?;
 
     let (client, backend, keystore_container, mut task_manager, on_demand) =
-        sc_service::new_light_parts::<VanillaBlock, VanillaRuntimeApi, Executor>(
+        sc_service::new_light_parts::<Block, RuntimeApi, Executor>(
             &config,
             telemetry.as_ref().map(|(_, telemetry)| telemetry.handle()),
         )?;
@@ -584,178 +458,4 @@ pub fn new_light(mut config: Configuration) -> Result<TaskManager, ServiceError>
     network_starter.start_network();
 
     Ok(task_manager)
-}
-
-/// Start a node with the given parachain `Configuration` and relay chain `Configuration`.
-///
-/// This is the actual implementation that is abstract over the executor and the runtime api.
-#[sc_tracing::logging::prefix_logs_with("Parachain")]
-async fn start_node_impl<RB>(
-    parachain_config: Configuration,
-    collator_key: CollatorPair,
-    polkadot_config: Configuration,
-    id: ParaId,
-    validator: bool,
-    rpc_ext_builder: RB,
-) -> sc_service::error::Result<(
-    TaskManager,
-    Arc<TFullClient<ParallelBlock, ParallelRuntimeApi, Executor>>,
-)>
-where
-    RB: Fn(
-            Arc<TFullClient<ParallelBlock, ParallelRuntimeApi, Executor>>,
-        ) -> jsonrpc_core::IoHandler<sc_rpc::Metadata>
-        + Send
-        + 'static,
-{
-    if matches!(parachain_config.role, Role::Light) {
-        return Err("Light client not supported!".into());
-    }
-
-    let parachain_config = prepare_node_config(parachain_config);
-
-    let params = new_partial_parallel(&parachain_config)?;
-    params
-        .inherent_data_providers
-        .register_provider(sp_timestamp::InherentDataProvider)
-        .unwrap();
-    let (mut telemetry, telemetry_worker_handle) = params.other;
-
-    let polkadot_full_node = cumulus_client_service::build_polkadot_full_node(
-        polkadot_config,
-        collator_key.clone(),
-        telemetry_worker_handle,
-    )
-    .map_err(|e| match e {
-        polkadot_service::Error::Sub(x) => x,
-        s => format!("{}", s).into(),
-    })?;
-
-    let client = params.client.clone();
-    let backend = params.backend.clone();
-    let block_announce_validator = build_block_announce_validator(
-        polkadot_full_node.client.clone(),
-        id,
-        Box::new(polkadot_full_node.network.clone()),
-        polkadot_full_node.backend.clone(),
-    );
-
-    let prometheus_registry = parachain_config.prometheus_registry().cloned();
-    let transaction_pool = params.transaction_pool.clone();
-    let mut task_manager = params.task_manager;
-    let import_queue = params.import_queue;
-    let (network, network_status_sinks, system_rpc_tx, start_network) =
-        sc_service::build_network(sc_service::BuildNetworkParams {
-            config: &parachain_config,
-            client: client.clone(),
-            transaction_pool: transaction_pool.clone(),
-            spawn_handle: task_manager.spawn_handle(),
-            import_queue,
-            on_demand: None,
-            block_announce_validator_builder: Some(Box::new(|_| block_announce_validator)),
-        })?;
-
-    let rpc_client = client.clone();
-    let rpc_extensions_builder = Box::new(move |_, _| rpc_ext_builder(rpc_client.clone()));
-
-    if parachain_config.offchain_worker.enabled {
-        sc_service::build_offchain_workers(
-            &parachain_config,
-            task_manager.spawn_handle(),
-            client.clone(),
-            network.clone(),
-        );
-    }
-
-    sc_service::spawn_tasks(sc_service::SpawnTasksParams {
-        on_demand: None,
-        remote_blockchain: None,
-        rpc_extensions_builder,
-        client: client.clone(),
-        transaction_pool: transaction_pool.clone(),
-        task_manager: &mut task_manager,
-        config: parachain_config,
-        keystore: params.keystore_container.sync_keystore(),
-        backend: backend.clone(),
-        network: network.clone(),
-        network_status_sinks,
-        system_rpc_tx,
-        telemetry: telemetry.as_mut(),
-    })?;
-
-    let announce_block = {
-        let network = network.clone();
-        Arc::new(move |hash, data| network.announce_block(hash, data))
-    };
-
-    if validator {
-        let proposer_factory = sc_basic_authorship::ProposerFactory::with_proof_recording(
-            task_manager.spawn_handle(),
-            client.clone(),
-            transaction_pool,
-            prometheus_registry.as_ref(),
-            telemetry.as_ref().map(|x| x.handle()),
-        );
-        let spawner = task_manager.spawn_handle();
-
-        let parachain_consensus = build_relay_chain_consensus(BuildRelayChainConsensusParams {
-            para_id: id,
-            proposer_factory,
-            inherent_data_providers: params.inherent_data_providers,
-            block_import: client.clone(),
-            relay_chain_client: polkadot_full_node.client.clone(),
-            relay_chain_backend: polkadot_full_node.backend.clone(),
-        });
-
-        let params = StartCollatorParams {
-            para_id: id,
-            block_status: client.clone(),
-            announce_block,
-            client: client.clone(),
-            task_manager: &mut task_manager,
-            collator_key,
-            relay_chain_full_node: polkadot_full_node,
-            spawner,
-            backend,
-            parachain_consensus,
-        };
-
-        start_collator(params).await?;
-    } else {
-        let params = StartFullNodeParams {
-            client: client.clone(),
-            announce_block,
-            task_manager: &mut task_manager,
-            para_id: id,
-            polkadot_full_node,
-        };
-
-        start_full_node(params)?;
-    }
-
-    start_network.start_network();
-
-    Ok((task_manager, client))
-}
-
-/// Start a normal parachain node.
-pub async fn start_node(
-    parachain_config: Configuration,
-    collator_key: CollatorPair,
-    polkadot_config: Configuration,
-    id: ParaId,
-    validator: bool,
-) -> sc_service::error::Result<(
-    TaskManager,
-    Arc<TFullClient<ParallelBlock, ParallelRuntimeApi, Executor>>,
-)> {
-    start_node_impl(
-        parachain_config,
-        collator_key,
-        polkadot_config,
-        id,
-        validator,
-        |_| Default::default(),
-    )
-    .await
 }
