+++ pallets/loans/Cargo.toml
@@ -2,30 +2,30 @@
 authors = ['Parallel Team']
 edition = '2018'
 name    = 'pallet-loans'
-version = '3.0.0'
+version = '1.0.0'
 
 [package.metadata.docs.rs]
 targets = ['x86_64-unknown-linux-gnu']
 
 [dependencies]
-codec             = { package = 'parity-scale-codec', version = '2.0.0', default-features = false }
-frame-support     = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', default-features = false, version = '3.0.0' }
-frame-system      = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', default-features = false, version = '3.0.0' }
-frame-benchmarking      = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', default-features = false, version = '3.0.0', optional = true }
-orml-currencies   = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
-orml-tokens       = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
-orml-traits       = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
-orml-utilities    = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
-pallet-balances   = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
-primitives        = { package = 'parallel-primitives', path = '../../primitives', default-features = false }
-serde             = { version = '1', features = ['derive'] }
-sp-runtime        = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
-sp-std            = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
+codec              = { package = 'parity-scale-codec', version = '2.0.0', default-features = false }
+frame-benchmarking = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', default-features = false, version = '3.0.0', optional = true }
+frame-support      = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', default-features = false, version = '3.0.0' }
+frame-system       = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', default-features = false, version = '3.0.0' }
+orml-currencies    = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
+orml-tokens        = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
+orml-traits        = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
+orml-utilities     = { git = 'https://github.com/open-web3-stack/open-runtime-module-library.git', default-features = false }
+pallet-balances    = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
+primitives         = { package = 'parallel-primitives', path = '../../primitives', default-features = false }
+serde              = { version = '1', features = ['derive'] }
+sp-runtime         = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
+sp-std             = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
 
 [dev-dependencies]
-sp-core = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
-sp-io   = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
 lazy_static = { version = "1.4.0" }
+sp-core     = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
+sp-io       = { git = 'https://github.com/paritytech/substrate.git', branch = 'rococo-v1', version = '3.0.0', default-features = false }
 
 [features]
 default = ['std']
@@ -45,3 +45,4 @@ std     = [
 ]
 
 runtime-benchmarks = ["frame-benchmarking"]
+
