+++ primitives/Cargo.toml
@@ -2,7 +2,7 @@
 authors = ['Parallel Team']
 edition = '2018'
 name    = 'parallel-primitives'
-version = '3.0.0'
+version = '1.0.0'
 
 [dependencies]
 codec      = { package = 'parity-scale-codec', version = '2.0.0', default-features = false }
